BOAC-5957, python API decorators to dedicated decorators.py
