+++ boac/api/degree_progress_student_controller.py
@@ -25,9 +25,10 @@ ENHANCEMENTS, OR MODIFICATIONS.
 
 from datetime import datetime
 
+from boac.api.decorators import can_edit_degree_progress, can_read_degree_progress
 from boac.api.degree_progress_api_utils import clone_degree_template, create_batch_degree_checks
 from boac.api.errors import BadRequestError, ResourceNotFoundError
-from boac.api.util import can_edit_degree_progress, can_read_degree_progress, normalize_accent_color
+from boac.api.util import normalize_accent_color
 from boac.externals.data_loch import get_basic_student_data, get_student_by_sid
 from boac.lib.http import tolerant_jsonify
 from boac.lib.util import get as get_param, is_int, to_bool_or_none, to_int_or_none
