+++ boac/api/student_controller.py
@@ -23,8 +23,9 @@ SOFTWARE AND ACCOMPANYING DOCUMENTATION, IF ANY, PROVIDED HEREUNDER IS PROVIDED
 ENHANCEMENTS, OR MODIFICATIONS.
 """
 
+from boac.api.decorators import advisor_or_peer_advisor_required, advisor_required
 from boac.api.errors import BadRequestError, ForbiddenRequestError, ResourceNotFoundError
-from boac.api.util import advisor_or_peer_advisor_required, advisor_required, is_unauthorized_domain, put_notifications
+from boac.api.util import is_unauthorized_domain, put_notifications
 from boac.externals.data_loch import get_admitted_students_by_sids, get_students_by_sids, match_students_by_name_or_sid
 from boac.lib.http import tolerant_jsonify
 from boac.lib.util import to_bool_or_none
